requirements.txt
Flask
pandas
flask-cors
python-dotenv
google-api-python-client
plotly

